ipchanged: support case where controllers are not gateway
Handler scripts for ipchanged do not work in deployments where gateway
for physical nodes and Neutron external network are external to the
controllers e.g. when using an external firewall. Refactor these so we
send gratuitous ARP to default gateway in all cases, and gratuitous ARP
to other pNodes in case when acting as gateway.